Consolidates the links and link helpers
These helpers were created at similar times in different PRs but should
be one file.
