+++ app/helpers/link_helper.rb
@@ -1,4 +1,12 @@
 module LinkHelper
+  def feed_link(base_path)
+    "#{base_path}.atom"
+  end
+
+  def print_link(base_path)
+    "#{base_path}/print"
+  end
+
   def govuk_styled_link(text, path: nil, inverse: false)
     return text if path.blank?
