+++ spec/helpers/link_helper_spec.rb
@@ -1,6 +1,18 @@
 RSpec.describe LinkHelper do
   include LinkHelper
 
+  describe "#feed_link" do
+    it "appends .atom to the base_path" do
+      expect(feed_link("/base_path")).to eq("/base_path.atom")
+    end
+  end
+
+  describe "#print_link" do
+    it "appends /print to the base_path" do
+      expect(print_link("/base_path")).to eq("/base_path/print")
+    end
+  end
+
   describe "#govuk_styled_link" do
     let(:text) { "Some text" }
     let(:path) { "/path" }
